+++ src/features/quiz/quizFlow/quizSolve/QuizFooter.tsx
@@ -2,7 +2,6 @@ import Button from '@/components/ui/Button';
 import TextButton from '@/components/ui/TextButton';
 import { useNavigate } from 'react-router-dom';
 import type { QuizFlowProps } from '@/types/quizView';
-import { completeMission } from '@/api/home/missionAPI';
 import clsx from 'clsx';
 
 // 퀴즈 하단부 (하단 버튼 영역 렌더링)
@@ -12,24 +11,11 @@ type QuizFooterProps = {
   selectedAnswer: QuizFlowProps['selectedAnswer'];
   onNext: () => void;
   onShowSummary?: () => void;
+  onGoHome: () => void;
 };
 
-const QuizFooter = ({ step, selectedAnswer, onNext, onShowSummary }: QuizFooterProps) => {
+const QuizFooter = ({ step, selectedAnswer, onNext, onShowSummary, onGoHome }: QuizFooterProps) => {
   const navigate = useNavigate();
-
-  const handleGoHome = async () => {
-    try {
-      const coinAmount = await completeMission('DAILY_QUIZ');
-      navigate('/', {
-        state: {
-          missionCompleted: true,
-          coinAmount,
-        },
-      });
-    } catch (error) {
-      console.log('미션 완료 실패: ', error);
-    }
-  };
 
   return (
     <div className='fixed bottom-0 left-0 right-0 flex justify-center pb-[20px]'>
@@ -53,14 +39,7 @@ const QuizFooter = ({ step, selectedAnswer, onNext, onShowSummary }: QuizFooterP
             <TextButton size='small' variant='default' fullWidth onClick={onShowSummary}>
               퀴즈 결과 확인하기
             </TextButton>
-            <Button
-              size='large'
-              variant='primary'
-              fullWidth
-              onClick={() => {
-                handleGoHome();
-              }}
-            >
+            <Button size='large' variant='primary' fullWidth onClick={onGoHome}>
               홈으로 이동하기
             </Button>
           </>
